chore!: remove ReactSelect
