+++ packages/core/upload/admin/src/components/SelectTree/SelectTree.jsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useMemo, useState } from 'react';
 
-/**
- * @note – This component is way too complex to convert to the DS version.
- */
-// eslint-disable-next-line no-restricted-imports
-import { ReactSelect as Select } from '@strapi/helper-plugin';
+import { Box } from '@strapi/design-system';
+import { Cross, CarretDown } from '@strapi/icons';
 import PropTypes from 'prop-types';
+import ReactSelect, { components } from 'react-select';
+import styled, { useTheme } from 'styled-components';
 
 import Option from './Option';
 import flattenTree from './utils/flattenTree';
@@ -68,6 +67,208 @@ const SelectTree = ({ options: defaultOptions, maxDisplayDepth, defaultValue, ..
       {...props}
     />
   );
+};
+
+const Select = ({ components, styles, error, ariaErrorMessage, ...props }) => {
+  const theme = useTheme();
+  const customStyles = getSelectStyles(theme, error);
+
+  return (
+    <ReactSelect
+      menuPosition="fixed"
+      components={{
+        ClearIndicator,
+        DropdownIndicator,
+        IndicatorSeparator: () => null,
+        LoadingIndicator: () => null,
+        ...components,
+      }}
+      aria-errormessage={error && ariaErrorMessage}
+      aria-invalid={!!error}
+      styles={{ ...customStyles, ...styles }}
+      {...props}
+    />
+  );
+};
+
+Select.defaultProps = {
+  ariaErrorMessage: '',
+  components: {},
+  error: undefined,
+  styles: {},
+};
+
+Select.propTypes = {
+  components: PropTypes.object,
+  styles: PropTypes.object,
+  error: PropTypes.string,
+  ariaErrorMessage: PropTypes.string,
+};
+
+const IconBox = styled(Box)`
+  background: transparent;
+  border: none;
+  position: relative;
+  z-index: 1;
+
+  svg {
+    height: ${11 / 16}rem;
+    width: ${11 / 16}rem;
+  }
+
+  svg path {
+    fill: ${({ theme }) => theme.colors.neutral600};
+  }
+`;
+
+const ClearIndicator = (props) => {
+  const Component = components.ClearIndicator;
+
+  return (
+    <Component {...props}>
+      <IconBox as="button" type="button">
+        <Cross />
+      </IconBox>
+    </Component>
+  );
+};
+
+const CarretBox = styled(IconBox)`
+  display: flex;
+  background: none;
+  border: none;
+
+  svg {
+    width: ${9 / 16}rem;
+  }
+`;
+
+const DropdownIndicator = ({ innerProps }) => {
+  return (
+    // @ts-expect-error – issue with the ref attached to `innerProps`
+    <CarretBox paddingRight={3} {...innerProps}>
+      <CarretDown />
+    </CarretBox>
+  );
+};
+
+DropdownIndicator.propTypes = {
+  innerProps: PropTypes.object.isRequired,
+};
+
+const getSelectStyles = (theme, error) => {
+  return {
+    clearIndicator: (base) => ({ ...base, padding: 0, paddingRight: theme.spaces[3] }),
+    container: (base) => ({
+      ...base,
+      background: theme.colors.neutral0,
+      lineHeight: 'normal',
+    }),
+    control(base, state) {
+      let borderColor = theme.colors.neutral200;
+      let boxShadowColor;
+      let backgroundColor;
+
+      if (state.isFocused) {
+        borderColor = theme.colors.primary600;
+        boxShadowColor = theme.colors.primary600;
+      } else if (error) {
+        borderColor = theme.colors.danger600;
+      }
+
+      if (state.isDisabled) {
+        backgroundColor = `${theme.colors.neutral150} !important`;
+      }
+
+      return {
+        ...base,
+        fontSize: theme.fontSizes[2],
+        height: 40,
+        border: `1px solid ${borderColor} !important`,
+        outline: 0,
+        backgroundColor,
+        borderRadius: theme.borderRadius,
+        boxShadow: boxShadowColor ? `${boxShadowColor} 0px 0px 0px 2px` : '',
+      };
+    },
+    indicatorsContainer: (base) => ({ ...base, padding: 0, paddingRight: theme.spaces[3] }),
+    input: (base) => ({
+      ...base,
+      margin: 0,
+      padding: 0,
+      color: theme.colors.neutral800,
+      gridTemplateColumns: '0 100%',
+    }),
+    menu(base) {
+      return {
+        ...base,
+        width: '100%',
+        marginTop: theme.spaces[1],
+        backgroundColor: theme.colors.neutral0,
+        color: theme.colors.neutral800,
+        borderRadius: theme.borderRadius,
+        border: `1px solid ${theme.colors.neutral200}`,
+        boxShadow: theme.shadows.tableShadow,
+        fontSize: theme.fontSizes[2],
+        zIndex: 2,
+      };
+    },
+    menuList: (base) => ({
+      ...base,
+      paddingLeft: theme.spaces[1],
+      paddingTop: theme.spaces[1],
+      paddingRight: theme.spaces[1],
+      paddingBottom: theme.spaces[1],
+    }),
+    menuPortal: (base) => ({
+      ...base,
+      zIndex: 100,
+    }),
+    option(base, state) {
+      let backgroundColor = base.backgroundColor;
+
+      if (state.isFocused || state.isSelected) {
+        backgroundColor = theme.colors.primary100;
+      }
+
+      return {
+        ...base,
+        color: theme.colors.neutral800,
+        lineHeight: theme.spaces[5],
+        backgroundColor,
+        borderRadius: theme.borderRadius,
+        '&:active': {
+          backgroundColor: theme.colors.primary100,
+        },
+      };
+    },
+    placeholder: (base) => ({
+      ...base,
+      color: theme.colors.neutral600,
+      marginLeft: 0,
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      whiteSpace: 'nowrap',
+      maxWidth: '80%',
+    }),
+    singleValue(base, state) {
+      let color = theme.colors.neutral800;
+
+      if (state.isDisabled) {
+        color = theme.colors.neutral600;
+      }
+
+      return { ...base, marginLeft: 0, color };
+    },
+    valueContainer: (base) => ({
+      ...base,
+      cursor: 'pointer',
+      padding: 0,
+      paddingLeft: theme.spaces[4],
+      marginLeft: 0,
+      marginRight: 0,
+    }),
+  };
 };
 
 const OptionShape = PropTypes.shape({
